moved get_data_model_from_task from util to sv_benchmarks_util
