+++ benchexec/tools/cbmc.py
@@ -8,7 +8,7 @@
 import logging
 from xml.etree import ElementTree
 
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 import benchexec.tools.template
 import benchexec.result as result
 
@@ -38,7 +38,7 @@ class Tool(benchexec.tools.template.BaseTool2):
         elif "--xml-ui" not in options:
             options = options + ["--xml-ui"]
 
-        data_model_param = util.get_data_model_from_task(
+        data_model_param = get_data_model_from_task(
             task, {"ILP32": "--32", "LP64": "--64"}
         )
         if data_model_param and data_model_param not in options:
