+++ benchexec/tools/esbmc.py
@@ -6,7 +6,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import os
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 import benchexec.tools.template
 import benchexec.result as result
 
@@ -32,9 +32,7 @@ class Tool(benchexec.tools.template.BaseTool2):
         return "ESBMC"
 
     def cmdline(self, executable, options, task, rlimits):
-        data_model_param = util.get_data_model_from_task(
-            task, {"ILP32": "32", "LP64": "64"}
-        )
+        data_model_param = get_data_model_from_task(task, {"ILP32": "32", "LP64": "64"})
         if data_model_param and "--arch" not in options:
             options += ["--arch", data_model_param]
         return (
