+++ benchexec/tools/fshell-witness2test.py
@@ -9,7 +9,7 @@ import re
 
 import benchexec.result as result
 import benchexec.tools.template
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 
 
 class Tool(benchexec.tools.template.BaseTool2):
@@ -63,7 +63,7 @@ class Tool(benchexec.tools.template.BaseTool2):
         if task.property_file:
             options = options + ["--propertyfile", task.property_file]
 
-        data_model_param = util.get_data_model_from_task(
+        data_model_param = get_data_model_from_task(
             task, {"ILP32": "-m32", "LP64": "-m64"}
         )
         if data_model_param and data_model_param not in options:
