+++ benchexec/tools/gacal.py
@@ -5,7 +5,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 import benchexec.tools.template
 import benchexec.result as result
 
@@ -35,9 +35,7 @@ class Tool(benchexec.tools.template.BaseTool2):
         return self._version_from_tool(executable)
 
     def cmdline(self, executable, options, task, rlimits):
-        data_model_param = util.get_data_model_from_task(
-            task, {"ILP32": "32", "LP64": "64"}
-        )
+        data_model_param = get_data_model_from_task(task, {"ILP32": "32", "LP64": "64"})
         if data_model_param and "--architecture" not in options:
             options += ["--architecture", data_model_param]
