+++ benchexec/tools/predatorhp.py
@@ -5,7 +5,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 import benchexec.tools.template
 import benchexec.result as result
 
@@ -34,7 +34,7 @@ class Tool(benchexec.tools.template.BaseTool2):
             else []
         )
 
-        data_model_param = util.get_data_model_from_task(
+        data_model_param = get_data_model_from_task(
             task,
             {"ILP32": "--compiler-options=-m32", "LP64": "--compiler-options=-m64"},
         )
