+++ benchexec/tools/two_ls.py
@@ -5,7 +5,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import benchexec.util as util
+from benchexec.tools.sv_benchmarks_util import get_data_model_from_task
 import benchexec.tools.template
 import benchexec.result as result
 
@@ -30,7 +30,7 @@ class Tool(benchexec.tools.template.BaseTool2):
         if task.property_file:
             options = options + ["--propertyfile", task.property_file]
 
-        data_model_param = util.get_data_model_from_task(
+        data_model_param = get_data_model_from_task(
             task, {"ILP32": "--32", "LP64": "--64"}
         )
         if data_model_param and data_model_param not in options:
